% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unpackingfunction.R
\name{RunTransmissionModel}
\alias{RunTransmissionModel}
\title{Run the transmission model}
\usage{
RunTransmissionModel(theta, tx_pars, runtime, stepsize, tx_times)
}
\arguments{
\item{theta}{named vector of model parameters; see description for details of elements}

\item{tx_pars}{named vector of model parameters relevant to treatment}

\item{runtime}{model run time, in years}

\item{stepsize}{model stepsize (dt) in years. Relatively small step sizes (1/12 or less) are recommended to avoid integration errors}

\item{tx_times}{a numeric vector of model treatment times (only used if input_tx_times==1 in tx_pars)}
}
\value{
A list containing the output from the simulated transmission model
}
\description{
This function allows the user to simulate an intestinal schistosomiasis transmission
model. See the package vignette for further details.
}
